requirements.txt
streamlit==1.33.0
pypdf==4.2.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
# Add the Google GenAI library
langchain-google-genai
# We still need sentence-transformers for the free embeddings
sentence-transformers==2.7.0
# Keep the stable LangChain stack
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1